pyproject.toml
[tool.poetry]
name = "givery-test"
version = "0.1.0"
description = ""
authors = ["Sebastian <ingesql1992@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
